requirements.txt
PyYAML==6.0.1
sigma-cli==1.0.0
